requirements.txt
aiohappyeyeballs
aiohttp
aiohttp-retry
aiosignal
attrs
beautifulsoup4
blinker
bs4
certifi
charset-normalizer
click
contourpy
cycler
Flask
fonttools
frozenlist
h11
idna
itsdangerous
Jinja2
joblib
kiwisolver
MarkupSafe
matplotlib
multidict
newsapi-python
numpy
outcome
packaging
pandas
pillow
plotly
propcache
PyJWT
pyparsing
PySocks
python-dateutil
python-dotenv
pytz
requests
scikit-learn
scipy
seaborn
selenium
six
sniffio
sortedcontainers
soupsieve
tenacity
threadpoolctl
trio
trio-websocket
twilio
typing_extensions
tzdata
urllib3
websocket-client
Werkzeug
wsproto
WTForms
yarl
